+++ projects/core/src/lib/pipes/find.pipe.spec.ts
@@ -1,0 +1,8 @@
+import { FindPipe } from './find.pipe';
+
+describe('FindPipe', (): void => {
+    it('create an instance', (): void => {
+        const pipe: FindPipe = new FindPipe();
+        void expect(pipe).toBeTruthy();
+    });
+});