+++ projects/core/src/lib/pipes/mosa-date.pipe.ts
@@ -4,10 +4,7 @@ import { TranslateService } from '@ngx-translate/core';
 import { BehaviorSubject, Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
 
-@Pipe({
-    name: 'mosaDate',
-    standalone: false,
-})
+@Pipe({ name: 'mosaDate' })
 export class MosaDatePipe implements PipeTransform {
 
     private readonly store$: BehaviorSubject<number>;
@@ -79,7 +76,7 @@ export class MosaDatePipe implements PipeTransform {
     }
 
     private template(key: string, val: number): string {
-        return this.templates[ key ] && this.templates[ key ].replace(/%d/i, String(Math.abs(Math.round(val))));
+        return this.templates[ key ]?.replace(/%d/i, String(Math.abs(Math.round(val))));
     }
 
     private updateTemplates(): void {