+++ projects/core/src/lib/pipes/mosa-duration.pipe.ts
@@ -2,12 +2,9 @@ import { Pipe, PipeTransform } from '@angular/core';
 import { TranslateService } from '@ngx-translate/core';
 import { BehaviorSubject, Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
-import { MosaDurationFormat } from '../../models/mosa-duration.model';
+import { MosaDurationFormat } from '../models/mosa-duration.model';
 
-@Pipe({
-    name: 'mosaDuration',
-    standalone: false,
-})
+@Pipe({ name: 'mosaDuration' })
 export class MosaDurationPipe implements PipeTransform {
 
     private readonly store$: BehaviorSubject<number>;