+++ projects/core/src/lib/utils/commons.util.ts
@@ -68,6 +68,15 @@ export function getTransformMatrix(element: HTMLElement): ITransformMatrix {
     };
 }
 
+export function some<T>(item: T, searchValue: T[keyof T] | T[keyof T][], searchKey?: keyof T): boolean {
+    let searchValues: T[keyof T][] = searchValue as T[keyof T][];
+    if (!Array.isArray(searchValue)) {
+        searchValues = [ searchValue ];
+    }
+
+    return searchValues.includes(searchKey ? item[ searchKey ] : item as T[keyof T]);
+}
+
 /**
  * Generates a random string with numbers and letters
  * @param length
@@ -89,7 +98,7 @@ export function tryJsonParse<T>(data: string, fallback?: () => T): T | null {
         if (!isNullOrEmpty(data)) {
             return JSON.parse(data);
         }
-    } catch (err: unknown) {
+    } catch {
         // Ignore - go on with fallback check
     }