+++ projects/material/src/lib/components/skeleton-loader/base.skeleton-loader.ts
@@ -4,7 +4,6 @@
     selector: 'mosa-skeleton-loader',
     template: '',
     changeDetection: ChangeDetectionStrategy.OnPush,
-    standalone: false,
 })
 // eslint-disable-next-line @angular-eslint/component-class-suffix
 export class BaseSkeletonLoader {